Task-3-Secure-File-Sharing-System/requirements.txt
Flask==3.0.0
pycryptodome==3.20.0
